Align listener variable name
